stock_project/requirements.txt
numpy
pandas
pandas_ta
beautifulsoup4
lxml
nasdaq-data-link
yfinance
yahooquery
python-dateutil
requests
matplotlib
seaborn
plotly
streamlit
st-annotated-text

